backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
openai==1.30.1
pytest==8.2.0
pytest-asyncio==0.23.7
websockets==12.0
python-dotenv==1.0.1
aiofiles==23.2.1
